pdf_reader/python3/requirements_python3.txt
tcex
pdfminer.six